pyproject.toml
[tool.poetry]
name = "cloudstaffhub"
version = "0.1.0"
description = "HR application"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
Django = "^5.0.1"
django-bootstrap-datepicker-plus = "^5.0.5"
django-bootstrap4 = "^23.4"
django-crispy-forms= "^2.1"
django-environ= "^0.11.2"
django-extensions= "3.2.3"
django-filter = "^23.5"
djangorestframework = "3.14.0"
django-phonenumber-field = "^7.3.0"
django-widget-tweaks = "^1.5.0"
factory-boy = "3.3.0"
Pillow = "10.1.0"
pre-commit = "^3.6.0"
psycopg2 = "^2.9.9"
crispy-bootstrap5 = "^2023.10"
phonenumbers = "^8.13.27"
funcy = "^2.0"
guardian = "^0.2.2"
django-guardian = "^2.4.0"
icecream = "^2.1.3"
safety = "^3.0.1"
bandit = "^1.7.7"
ruff = "^0.1.14"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
